Limit search to 500 transactions
When you do a search there is a parameter to specify the number of
results returned called display_size. Currently this defaults to 500
but does not have a max value. As we have seen the speed of searches is
very dependent on the number of results returned we want limit how big
this can be. Have limited display_size to be 500 or less.