PP-3733: Direct debit payment integration test with pact
The way of building pacts is using the java dsl, see
https://github.com/DiUS/pact-jvm/tree/master/pact-jvm-consumer-junit#the-pact-junit-dsl.
However doing it this way is really clunky and get unmanageable and unreadable
for larger json bodies. It just seems like defining the json pact json file and
using it would it more manageable and easier to read, and would make our java
test classes more succint as it won't be cluttered with the pact dsl setups.

This commit introduces a PactProviderRule which enables us to annotate each
junit test with:

```
@Pacts(pacts = {"publicapi-publicauth", "publicapi-direct-debit-connector"})
```

This indicates that the publicapi-publicauth.json and
publicapi-direct-debit-connector.json files, which represent downstream
dependencies of publiapi, will be used to set up the pact mock servers.

@oswaldquek