Inline all resource urls and field name constants
The url paths are almost exclusively used only in one place, which is the
configuration of the resource endpoint. It's much clearer to have the string
literal visible on the resource annotation than to have it obscured by
constant.

We also inlined some constants which define the keys from request payloads.
These are also part of our external API, not likely to change and used only in
one or two places. Again, having a string literal makes the code easier to read
in these cases.

With @heathd