PP-3842: Fix consumer pact
While running the publicapi-direct-debit-connector pact for real against the
direct-debit-connector service for real, I noticed some of the interactions
defined here are incorrect. For example, in the call to
`/v1/api/accounts/GATEWAY_ACCOUNT_ID/charges` (on the direct debit connector),
the email and payment provider are not returned in the response body.

Also add a provider state and correct [regex] matchers where appropriate as
some of the data is dynamic. For example, the payment requested external id and
charge token id (UUID) are generated.

@oswaldquek