PP-4005 Added AuditRate to RateFilter
Adds new `auditRateOf` method to rate filter, to allow us to see the effect
of a proposed rate limit without imposing it. This is useful to understand
the behaviour of our current rate limit set up, which works per service per node,
and therefore is abit unpredictable in its effect.

@kbottla  with @mrlumbu