PP-4005 Segment rate limit by method
In order to more precisely rate limit the different parts of
public api, this commit passes a concatenation of method and
api key as rate limiting key. This will mean GET requests cannot impinge
on POSTs, so if a service does a batch of GETs it won't block the
mor eimportant process of creating new payments. This is a small
change in right direction, rather than a total solution.

We may want to set different rates for POSTs and GETs at some point,
but this is not currently required.