bau: Use wiremock
I found it difficult to get upgrade the mockserver library
(https://github.com/alphagov/pay-publicapi/pull/435) as there's a concept of a
mockserver client and server. Conceptually wiremock is much simpler as we're
only mocking out the services that publicapi connects to (e.g. connector,
publicauth).

I've had to set the rateLimiter.noOfReq and rateLimiter.noOfReqForPost to 1000
as a fair number of classes extend PaymentResourceITestBase which has a static
API_KEY, which meant the rate limit would have been hit causing many tests to
fail.