PP-5224 Split up CreatePaymentRequest
CreatePaymentRequest is currently used to represent both Card and DirectDebit payment
requests. This is becoming confusing and impractical as validation requirements
diverge between the two. This commit create dedicated sub-type for each.
1. Create CreatCardPaymentRequest and CreateDirectDebitRequest classes. Each
implementing the attributes and validation require beyond the base class of the
existing CreatePaymentRequest.
2. Simplify ViolationExceptionMapper to convert to the api field name using a
camel case to snake case conversion. This seems safe as it follows our convention
and avoids the complexity of reflection which would need extending to accomodate
the newly introduced hierachy (if the field is not found on the sub-class then
it would have to search the super class).
3. Remove validation from PaymentResource method and into the new CreateDirect-
DebitPaymentRequest.
4. Reinstate ignored test now that a request can have either return_url or
agreement_id and validate independently.