PP-5242 Return errors if agreement_id or return_url missing
We determine whether a create payment request is a card or direct debit
request when we deserialise the JSON (using RequestJsonParser.js). At
this point we do not know whether the account is a card or direct debit
account, so we look at whether the request contains a return_url or an
agreement_id.

Previously, if there was no agreement_id or return_url, the error would
always state that the return_url was missing. However, this is only true
for card payments, so the error now states that one or the other of
return_url or agreement_id is required.

Additionally, there was no validation done to check that the correct
form of the request was made for the correct account type (card/dd).
Now, if there is a mismatch an error will be returned stating either
that the return_url or agreement_id is missing, as appropriate.