PP-5301 Remove superfluous search classes
The SearchCardPayments, SearchDirectDebitPayments, SearchRefunds classes
existed because the service for searching payments didn't know what type
of payment it was searching and obtained these using a factory.

However, this is no longer necessary as the endpoints have been split
out and they now use different services. Move the code from these
classes into the services and move the shared code in the base class for
adding pagination links to a new util class, PaginationDecorator.

Also, remove parameter validation methods as these only achieved
the purpose of validating whether search parameters had been provided
that were unsupported for the payment method. Now that the endpoints
have been separated this is unnecessary.