PP-5299: Temporarily remove links expectation when getting a mandate
In an ideal scenario, removing the next_url and next_url_post links would work,
even though direct-debit-connector still the returns the links currently.
However it does not as Postel's Law only seems to be followed for json keys and not
for json values - an array of links in this case.

As a reminder of what
Postel's Law is, from https://docs.pact.io/getting_started/matching/gotchas:

```
Be liberal in what you accept - when verifying a pact in the provider project,
the response body and headers may contain fields that were not defined in the
expectations, on the assumption that any extra field will be ignored by your
consumer. This allows a provider to evolve without breaking existing consumers
```

Therefore the way to get all this working is:
- Temporarily remove the links from the pact (this commit). This will enable
  this version of publicapi to be deployed through to production even though
  dd-connector is still returning the next_url and next_url_post links in
  production.
- Amend dd-connector to not return next_url and next_url_post and deploy that
  through to production.
- Put the links back into publicapi's pact. The only link that will exist is
  the "self" and deploy that to production.