PP-5612 Add gateway account id to logs
To aid debugging this adds the gateway account id to the MDC using a
`ContainerRequestFilter` so it can be logged in subsequent lines. The value is
cleared by  a new `ClearMdcValuesFilter` class which is a
`ServletFilter` at the head of the filter chain, this allows it to clean the
mdc values within a `finally` block after calling `chain.doFilter()`.

An alternative to using the `ClearMdcValuesFilter` to clean up at the
end of the request would be follow the pattern of the `LoggingFilter`
which we use to add the request-id. This will replace the request-id if
it already exists, else set it a new. Although this seems simpler it
has the risk of any log lines printed prior to `LoggingFilter`
containing a stale request-id; instead it seems safer to clean the mdc
values once the request has completed and the thread released to the
pool.

Ultimately it seems dealing with MDC and threadpools is a little tricky
but we think this is a reasonable solution.