PP-5761: Log X-Request-Id temporarily
It seems when the x-request-id gets into connector, it's appended with a comma.
Search `atlkbga69vtvp2kk64lnm2v7ae` in the test environment in splunk to see
what I mean. As this doesn't happen for calls from frontend to connector, and
https://github.com/alphagov/pay-publicapi/pull/779 was the only thing that was
deployed, log the headers to see if anything is actually happening here.