BAU Pull redis docker image prior to CI build
The docker registry v1 API, which is used by the spotify docker-client
maven plugin that we use has been switched off.
https://www.docker.com/blog/registry-v1-api-deprecation/

We need to use the v2 API, which test containers use.

In the meantime, pull the docker image prior to running tests on CI so
the build will pass