PP-6010 Add Pact test for searching ledger with a non existent page no
Add a pact test to verify that ledger returns a 404 when a search
request is made with a page number that does not exist.

This will allow us to remove an end-to-end test checking this as we
already have an integration test to verify that we handle a 404 response
from ledger correctly.