Change URLValidator implementation to avoid initialization exception
Was getting

```
java.lang.ExceptionInInitializerError
	at uk.gov.pay.api.validation.URLValidatorTest.whenUrlIsBlank_shouldFailValidation_whenEnabledSecureConnection(URLValidatorTest.java:98)
Caused by: java.lang.IllegalStateException: Can only invoke this method before calling getInstance
	at uk.gov.pay.api.validation.URLValidatorTest.whenUrlIsBlank_shouldFailValidation_whenEnabledSecureConnection(URLValidatorTest.java:98)
```

something has changed in the update of commons-validator from 1.6 to 1.7 which likely initializes `UrlValidator` and therefore causes the static initializer segment to not run on time.

The order of execution of static blocks in enums is non-intuitve. In fact the enum instances `SECURITY_ENABLED` and `SECURITY_DISABLED` are initialzed first before the static initializer runs.

Refactor the implementation to avoid using `DomainValidator.updateTLDOverride` (which operats on the global/static level) and instead to pass in a `DomainValidator` instance with the appropriate overrides applied when the `UrlValidator` instance is created.

This seems to make the tests pass