PP-7242 Use regex from CardExpiryDate to validate card expiry dates
Replace the current regex for validating card expiry dates with the
one from CardExpiryDate from Pay Java commons.

The two regexes are not identical. The old one is:

(0[1-9]|1[0-2])/\d{2}

… while the new one is:

(0[1-9]|1[0-2])/([0-9]{2})

However, these are equivalent.

Both regexes are identical up to and including the slash. After
this, the new regex has parentheses (used for a capturing group)
but these don’t actually make any difference to what’s matched.

In Java, the character class \d matches only the digits in the
range 0 to 9 inclusive, so \d{2} and [0-9]{2} are equivalent (this
is not the case in all languages!).