PP-7104 Log when API keys are used
Log at INFO level when a valid API key is used. Log the token_link for
the API key, which is returned by public auth.

Log at WARN level when an API key with an invalid checksum is used. Log
the IP address the request originates from to help us detect brute force
attempts.

Log at WARN level when an invalid API key with a valid checksum is used.

Log at WARN level when a revoked API key is used. Log the token_link for
the API key, which is returned by public auth.

Remove logic that maps a non-existent/unknown value for `token_type`
returned from public auth to TokenPaymentType.CARD. Public auth will
always return a `token_type`, and if it doesn't this will result in a
`com.fasterxml.jackson.databind.exc.InvalidFormatException` which will
be wrapped in a 500 response. This will show up in Sentry so we should
be able to catch it if there are any future issues introduced.

Add an integration test to ensure this is the case.