PP-7560 Clear MDC at both the beginning and end of a request
We are observing in our logs that sometimes the remote_address
from a previous request seems to be logged even though the MDC
should have been cleared at the end of that request.

Additionally clear the MDC at the beginning of the request (it
should have no values at this point so we shouldn’t lose anything
we want) to see if this makes things better.

with @katstevens