PP-8095 Make security.txt require cache revalidation every request
Send a Cache-Control: no-cache header and an Expires header with
the current time when redirecting from /.well-known/security.txt
and /security.txt.

Without these headers, clients will use heuristics to decide how
long to cache the response for, which could be a problem if we
want to update our security.txt.