PP-8415 Stop logging payment reference
The payment reference can come from user input for a payment link
payment. We've seen users mistakenly entering their card number in the
reference field.

It is a lot of effort to remove log lines from all places when this
happens, so stop logging the reference in all cases. We can still find
out the reference or search for payments by the reference in toolbox.