PP-6699: Remove Direct Debit code part I
We are no longer running direct debit apps and infra (such as database, task
defs) have been removed too.

The code removed in this commit makes the build pass as a result of removing
the `connectorDDUrl` property in the config.yaml. There are more classes to be
removed which will come in a later commit.