PP-8844 Add timer metric to redis
creates two timer metrics:
- one around the incr operation
- one around the expire operation

this should accumulate and emit a counter and latency distribution metrics for
these two operations, as percieved by the redis client

Tests updated to inject a mock environment, metrics registry etc. We use a concrete `Timer()` object because it's easier than mocking it and it does very little when it's not hooked up to the metrics registry.

We used the `time()` method on timer which invokes a callable and can raise an exception.