PP-9142 Accept authorisation_mode and return in responses.
Accept a new authorisation_mode property when creating a new payment.
Validate that this is a valid AuthorisationMode, and is one of the
allowed values - "web" or "moto_api". Ensure payments cannot be created
with `authorisation_mode` of "external", as this should only be set
internally for payments created using the telephone payment notification
API.

Return the `authorisation_mode` from connector/ledger in get and search
responses.

In the swagger file, hide the property on the create payment request as
we do not want people creating payments with this property set yet.
However, document the property on the payment response objects, as we
will start returning it immediately with a value of "web" for existing
payments.