PP-9495: add description and user_identifier to create agreement
Accept and validation `description` and `user_identifier` fields on the
create agreement endpoint.

`description` is required and will throw an appropriate error if
missing.

`user_identifier` is optional but must be a string if set.

Note this does not extend the expected agreement repsonse from Connector
as Ledger will be used to serve agreements and this should be removed in
the next change.