PP-9407 Don't require Authorisation header for auth API
The MOTO payments authorisation API should not require an Authorisation
header with the bearer token to be present. We instead require a
one_time_token in the request body to authorise the request.

Add an exclusion to the filter that checks for Authorisation header for
the URL "/v1/auth". It is not possible to add this exclusion in the
URL pattern we provide for the filter when registering it in the app
module, so instead hardcode this exclusion into the filter.