PP-9389 Use connector return_url validation
Make the return_url no longer mandatory as it is not required for
payments with `authorisation_mode` of `moto_api`. Instead, just validate
that it is a String if it is present. Validation that the return_url is
present and valid if required is now done by connector, and returns error
responses that we can map in public api.

Map the error_identifiers MISSING_MANDATORY_ATTRIBUTE,
INVALID_ATTRIBUTE_VALUE and UNEXPECTED_ATTRIBUTE returned by connector
when creating a payment to pass on the validation error message from
connector along with the appropriate HTTP status and error code.

Add pact tests for the connector validation responses.