PP-9812 Handle connector response for recurring payment request
When public API is handling a recurring payment request
("authorisation_mode": "agreement" and an "agreement_id"), we
need to handle a couple of things differently depending on the
response from connector.

If the payment is created successfully, we need to not include a
cancel link in the public API response.

If connector responds with an error with the error identifier
AGREEMENT_NOT_FOUND, public API needs to return a 400 response
like this:

{
  "field": "agreement_id",
  "code": "P0102",
  "description": "Invalid attribute value: agreement_id. Agreement does not exist"
}

If connector responds with an error with the error identifier
AGREEMENT_NOT_ACTIVE, public API needs to return a 400 response
like this:

{
  "field": "agreement_id",
  "code": "P0102",
  "description": "Invalid attribute value: agreement_id. Agreement must be active"
}

with @sfount