PP-9545: separate the ledger response and API model for agreeements
Splits out the POJO used to deserialise the response from Ledger and the
model used to respond in the API.

There are a number of internal properties and properties that will be
renamed. Separate models are used to keep this separation resilient to
future changes (vs. annotating accessors separately and hiding
properties).