PP-9545 Assert that we do not send "payment_instrument": null
It’s possible for an agreement to not have a payment instrument
yet. We have unintentionally flip-flopped between including
"payment_instrument": null or nothing in the agreement response
if this is the case.

We have now decided to not include "payment_instrument": null when
there is no payment instrument for the agreement. The rationale for
including it was it alerts services that there may sometimes be a
payment instrument. However, we wish to use the agreements model
for other kinds of recurring payments in the future and our crystal
balls are not good enough to rule out the possibility that we may
one day have a kind of agreement that never has a payment
instrument. If this happens, always including "payment_instrument":
null will be undesirable.

Update integration tests to ensure we keep this desired behaviour
of agreements only including "payment_instrument" if there is a
payment instrument.