Make PaymentWithAllLinks extend CardPayment
Given CardPayment is under the uk.gov.pay.api.model package, having links is always going to be part of the API response for a card, and we are never going to use CardPayment as a database model, so we can simply make PaymentWithAllLinks extend CardPayment. This means we can get rid of the @JsonUnwrapped feature which makes the code cleaner and easier to understand.