PP-11237: Remove redundant classes (#1787)
* PP-11237: Remove redundant classes

Payment can be deleted. We used to have a direct debit payment model, hence
this super class but this is no longer needed.

GetPaymentResult - nothing special about this class, just extends CardPayment,
so can be deleted and replaced with PaymentWithAllLinks.

Also make PaymentWithAllLinks extend CardPayment. Given CardPayment is under the uk.gov.pay.api.model package, having links is always going to be part of the API response for a card, and we are never going to use CardPayment as a database model, so we can simply make PaymentWithAllLinks extend CardPayment. This means we can get rid of the @JsonUnwrapped feature which makes the code cleaner and easier to understand.