bau: Test and PaymentWithAllLinks constructor cleanup (#1793)
* bau: Test and PaymentWithAllLinks constructor cleanup

Removed JsonAssert usages in PaymentsResourceSearchIT as they weren't doing
anything. Try adding the following in
`PaymentsResourceSearchIT.searchPayments_shouldOnlyReturnAllowedProperties` in
the current master branch:
```
JsonAssert.with(responseBody).assertNotDefined("_links.refunds.href")
```

It should cause the test to fail but it doesn't. Am not convinced that class does what we think it's doing.

Also use a builder pattern for PaymentWithAllLinks.