PP-12146 Create payment in payments microservice
For a POST create payment request, create the payment in the prototype
payments microservice rather than connector.

Note that the payments microservice currently only consumes and stores
basic payment attributes and not the full set of possible attributes
e.g. language, metadata.

Not yet done anything about determining the status of the payment -
hardcoding it to "created" in the response.