ZD-5631868 Add payment_id to refund OpenAPI spec
The payment_id field is included in the refund object and
documented at
https://docs.payments.service.gov.uk/api_reference/search_refunds_reference/#attributes-you-ll-get-in-a-search-refunds-response
but it is not included in our OpenAPI spec.

Since it’s already there and documented, people will be relying on
it so the only sensible thing to do is to include it in the OpenAPI
spec.

Description nabbed from the documentation and is identical to that
used elsewhere in the OpenAPI spec.