PP-13429: Clarify rate limiting functionality (#2178)
The rate limiting implementation can be quite to understand; hopefully adding
the comments will help. Rate limiting is currently applied using discrete
time blocks, but an idea for the future is to use a sliding window and do a
count for requests in the last second/minute/hour window which could
potentially make the logic simpler.

This commit also makes the RateLimiterKey constructor private; this is more
correct as it's currently public for the sake of convenient unit testing.