PP-14054 Update tests given new validation
With the previous commit, we have introduced validation for the description
and reference fields in order to block use of illegal characters which might
be used for XSS attacks.

With this change, we are now updating tests that were using placeholders
containing illegal characters, using valid values instead.

Additionally, we are also adding a couple of tests to ensure that values with
illegal characters are actually duly blocked.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-14054